+++ gradle/libs.versions.toml
@@ -1,30 +1,30 @@
 [versions]
-activityCompose = "1.9.0"
-coilCompose = "2.5.0"
-composeBom = "2024.06.00"
+activityCompose = "1.9.2"
+coilCompose = "2.7.0"
+composeBom = "2024.09.02"
 constraintlayout = "2.1.4"
 core = "3.3.0"
 coreKtx = "1.13.1"
 coreVersion = "1.6.1"
-fragmentKtx = "1.8.0"
-gradle = "8.5.0"
-gson = "2.10.1"
+fragmentKtx = "1.8.3"
+gradle = "8.5.2"
+gson = "2.11.0"
 hiltAndroid = "2.51.1"
 hiltNavigationCompose = "1.2.0"
 junit = "4.13.2"
 kotlinStdlibJdk7 = "2.0.0"
-kotlinxCoroutinesAndroid = "1.8.0"
+kotlinxCoroutinesAndroid = "1.8.1"
 kotlinxCoroutinesTest = "1.6.2"
-lifecycleLivedataKtx = "2.8.2"
-loggingInterceptor = "4.11.0"
+lifecycleLivedataKtx = "2.8.6"
+loggingInterceptor = "4.12.0"
 material = "1.12.0"
 mockitoKotlin = "5.2.1"
-navigationCompose = "2.7.7"
-navigationFragmentKtx = "2.7.7"
-navigationUiKtx = "2.7.7"
-pagingRuntimeKtx = "3.3.0"
+navigationCompose = "2.8.1"
+navigationFragmentKtx = "2.8.1"
+navigationUiKtx = "2.8.1"
+pagingRuntimeKtx = "3.3.2"
 picasso = "2.71828"
-retrofit = "2.9.0"
+retrofit = "2.11.0"
 robolectric = "4.12"
 roomRuntime = "2.6.1"
 symbolProcessingApi = "2.0.0-1.0.21"
